feature: GameSuite covers GameState test-only

+++ src/main/scala-2.11/nl/amsscala/simplegame/SimpleCanvasGame.scala
@@ -7,7 +7,7 @@ import scala.scalajs.js.JSApp
  * Main entry point for application start
  */
 object SimpleCanvasGame extends JSApp with Game with Page {
-  type T = Long // This sets the generic used by the whole application and tests.
+  type T = Int // This sets the generic used by the whole application and tests.
 
   /**
    * Entry point of execution
